# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/notifications/debug/send-test": {
      "post": {
        "operationId": "NotificationsController_sendTestNotification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTestNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebugNotificationResponseDto"
                }
              }
            }
          }
        },
        "summary": "DEBUG: Send a test notification to the current user\nPOST /notifications/debug/send-test\nOnly available to OWNER users",
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/api/push/subscribe": {
      "post": {
        "operationId": "PushNotificationsController_subscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribePushDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageDto"
                }
              }
            }
          }
        },
        "summary": "Subscribe to push notifications",
        "tags": [
          "PushNotifications"
        ]
      }
    },
    "/api/push/unsubscribe": {
      "delete": {
        "operationId": "PushNotificationsController_unsubscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnsubscribePushDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageDto"
                }
              }
            }
          }
        },
        "summary": "Unsubscribe from push notifications",
        "tags": [
          "PushNotifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kraken API",
    "description": "Kraken Chat Application API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "NotificationAuthorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "displayName",
          "avatarUrl"
        ]
      },
      "NotificationMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "content"
        ]
      },
      "NotificationDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "USER_MENTION",
              "SPECIAL_MENTION",
              "DIRECT_MESSAGE",
              "CHANNEL_MESSAGE",
              "THREAD_REPLY"
            ]
          },
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "messageId": {
            "type": "string",
            "nullable": true
          },
          "channelId": {
            "type": "string",
            "nullable": true
          },
          "directMessageGroupId": {
            "type": "string",
            "nullable": true
          },
          "authorId": {
            "type": "string"
          },
          "parentMessageId": {
            "type": "string",
            "nullable": true
          },
          "read": {
            "type": "boolean"
          },
          "dismissed": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/NotificationAuthorDto"
          },
          "message": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/NotificationMessageDto"
              }
            ]
          }
        },
        "required": [
          "type",
          "id",
          "userId",
          "messageId",
          "channelId",
          "directMessageGroupId",
          "authorId",
          "parentMessageId",
          "read",
          "dismissed",
          "createdAt"
        ]
      },
...
      "SendTestNotificationDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "USER_MENTION",
              "SPECIAL_MENTION",
              "DIRECT_MESSAGE",
              "CHANNEL_MESSAGE",
              "THREAD_REPLY"
            ]
          },
          "customMessage": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ]
      },
      "DebugNotificationResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "notification": {
            "$ref": "#/components/schemas/NotificationDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "notification",
          "message"
        ]
      },
...
      "PushSubscriptionKeys": {
        "type": "object",
        "properties": {
          "p256dh": {
            "type": "string"
          },
          "auth": {
            "type": "string"
          }
        },
        "required": [
          "p256dh",
          "auth"
        ]
      },
      "SubscribePushDto": {
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "keys": {
            "$ref": "#/components/schemas/PushSubscriptionKeys"
          },
          "userAgent": {
            "type": "string"
          }
        },
        "required": [
          "endpoint",
          "keys"
        ]
      },
      "SuccessMessageDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
      "UnsubscribePushDto": {
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string"
          }
        },
        "required": [
          "endpoint"
        ]
      },
...
    }
  }
}